Update add-custom-domain-name.md
There are two subsequent steps required to add a custom-domain name, requiring two distinct entries to DNS - which could be done in one go - but which are in different documentations.
As such customers will first request the TXT to be entered - and later they must again request a CNAME or several A records to be entered. Often this is a lot of overhead.

Here we add a note to do check the nextdocumenation on CNAME and A Records as well to point to this fact.