Update latest.md
Established a convention for stable annotation of edited changes by removing a unique token (`***-***`) generated by the changelog script. Commented lines are recognized as having been suppressed from the release notes, to allow for copy-pasting the entire block for reference when using the changelog script to add new changes after a merge to the release candidate branch.