Update current.md
From Shweta Dua, June 12, 2024 - We have Real Use Monitoring section on two Release notes – Cloud Manager and Cloud service and it is confusing as Real Use monitoring is now GA but reflecting different information on two different release notes.

Can we remove this section from Cloud manager (Early Adopter Programs) existing release notes and going forward? Real Use monitoring is now GA and it would be great to remove it as soon as possible as it is confusing to our customers and internal.