Update release-notes-current.md
added a new section in Sites to announce upcoming dates for setting the new Content Fragment Editor as default in AEMaaCS environments 

**Content Fragment Editor Rollout**

We will continue enabling the new Spectrum UI-based Content Fragment Editor in AEM as a Cloud Service. After becoming the default for all Cloud Service Developer environments in November 2024, it will be set as default for all Stage environments on April 1, 2025, and for all Production environments on May 1, 2025. In all cases, users will still have the option to revert to the traditional Content Fragment Editor in AEM Touch UI.